Count a cell's age buckets in u32, and hold a file to its own header
The Recency column was eight `u64` counts, 64 of a cell record's 230 bytes —
better than a quarter of `index.bin`, in the file every client downloads and
every walk reads, for a column nothing reads back yet.

`u32` is exact and half the width. A bucket counts systems, a cell's buckets
sum to its `count`, and the root's count is the galaxy's: 129 million against
the four billion a `u32` holds, so there are five doublings of headroom over
every system on record. Measured on the built index, `Aggregate` goes 204 to
172 bytes, `Cell` 230 to 198, and `index.bin` comes out 13.9% smaller.

`galaxy.md` budgets these at sixteen bytes by storing each bucket as a `u16`
share of `count` instead. Measured against a real build that is the wrong
trade: the worst absolute error is 17 systems, but six of 8,855 non-empty
buckets round away to nothing outright — and the bucket that rounds away is
the small one, which is the recently-changed one the axis exists to show. It
gets worse with scale, a cell at 129 million systems being some sixty times
the one measured. Exact at twice the width is the better answer, and the
field's doc now says so rather than leaving the doc's figure to be read as
the target.

`INDEX_VERSION` stays at zero, the format being explicitly unfrozen and a
rebuild being the fix.

That turned out to need a second change. A stale index carries the same magic
and the same zero, so nothing in the header tells it apart — and the decoder
read `count` records of whatever width it was compiled for, from a file
written at another, and handed back a plausible tree of nonsense. Silently: I
checked, by decoding a file padded to the old width, and got an index rather
than a refusal. So a width change at version zero was only safe by accident.

It is safe on purpose now. The header says how many cells follow and a cell is
a fixed width, so the body's length is a thing the file can be held to, and
anything but exactly `count * Cell::LEN` is refused. A rebuild instead of a
wrong sky. `an_index_of_another_width_is_refused` holds both directions, since
a record may grow as easily as shrink, and fails against the old decoder with
"a wider record read".

The payload keeps its tolerance for a short tail: a payload block carries no
count to be held against.

+++ galos_index/src/aggregate.rs
@@ -87,8 +87,18 @@ pub struct Aggregate {
     /// Counts per age bucket, a column of the record so a Recency span can be
     /// answered by prefix sum off the aggregates alone. Every build writes it
     /// and nothing reads it back yet: the Recency filter asks per system, off
-    /// `updated_at`. It stays because dropping it changes the record width.
-    aged: [u64; AGE_BUCKETS],
+    /// `updated_at`.
+    ///
+    /// `u32` rather than `u64`, which halves what the column costs and loses
+    /// nothing: a bucket counts systems, the buckets of one cell sum to its
+    /// `count`, and the root's count is the galaxy's — 129 million against the
+    /// four billion a `u32` holds, so there are five doublings of headroom
+    /// over every system on record. `galaxy.md` budgets these at sixteen bytes
+    /// by storing each bucket as a `u16` share of `count` instead; measured
+    /// over a real build that rounds the smallest buckets away to nothing, and
+    /// the smallest bucket is the recently-changed one the axis exists to
+    /// show. Exact at twice the width is the better trade.
+    aged: [u32; AGE_BUCKETS],
 }
 
 impl Aggregate {
@@ -262,7 +272,7 @@ record! {
         flux: [f64; TEMP_BUCKETS],
         light: Moments,
         mass: Moments,
-        aged: [u64; AGE_BUCKETS],
+        aged: [u32; AGE_BUCKETS],
     }
 }
